scraper to view youtube videos through youtube api
Using youtube's google api, we are scraping youtube (v3 of the api) to scrape videos getting 1. the title 2. the description 3. the tags on the videos. sadly, shorts do not have tags.

next step is getting the scraper to be modified so that the next step which is the brainrot detector boolean to work by looking at this data and reading it properly